refactor(lib): move request_handlers
