+++ app.js
@@ -26,7 +26,7 @@ const swaggerJSDoc = require('swagger-jsdoc')
 const chalk = require('chalk')
 const passport = require('passport')
 const controllers = require('./app/controllers')
-const requestHandlers = require('./lib/request_handlers')
+const requestHandlers = require('./app/lib/request_handlers')
 const initCloudinary = require('./app/lib/cloudinary')
 const compileSVGSprites = require('./app/lib/svg-sprite')
 const appURL = require('./app/lib/url')
